changes to SelectAll - executes stored procedure
The SelectAll procedure of the gateway class now executes a
stored procedure. This is currently a workaround for the current
development effort of a different project. The gatway classes
will eventually call stored procedures for crud statements.
This is all part of an effort to remove the ties to the
Microsoft.Enterprise classes.